Refactor events to decouple k8s event and cloud event
This commit refactors the code in events to decouple the k8s events emit
and cloud events emit. This commit fixes #4404.
