+++ pkg/reconciler/events/cloudevent/cloud_event_controller_test.go
@@ -22,11 +22,12 @@ import (
 	"time"
 
 	"github.com/google/go-cmp/cmp"
+	"github.com/tektoncd/pipeline/pkg/apis/config"
 	"github.com/tektoncd/pipeline/pkg/apis/pipeline/v1alpha1"
 	"github.com/tektoncd/pipeline/pkg/apis/pipeline/v1beta1"
 	resourcev1alpha1 "github.com/tektoncd/pipeline/pkg/apis/resource/v1alpha1"
+	"github.com/tektoncd/pipeline/pkg/reconciler/events/k8sevent"
 	"github.com/tektoncd/pipeline/test/diff"
-	eventstest "github.com/tektoncd/pipeline/test/events"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/tools/record"
@@ -621,7 +622,7 @@ func TestSendCloudEventWithRetries(t *testing.T) {
 			ceClient := Get(ctx).(FakeClient)
 			ceClient.CheckCloudEventsUnordered(t, tc.name, tc.wantCEvents)
 			recorder := controller.GetEventRecorder(ctx).(*record.FakeRecorder)
-			if err := eventstest.CheckEventsOrdered(t, recorder.Events, tc.name, tc.wantEvents); err != nil {
+			if err := k8sevent.CheckEventsOrdered(t, recorder.Events, tc.name, tc.wantEvents); err != nil {
 				t.Fatalf(err.Error())
 			}
 		})
@@ -675,6 +676,101 @@ func TestSendCloudEventWithRetriesNoClient(t *testing.T) {
 	if d := cmp.Diff("No cloud events client found in the context", err.Error()); d != "" {
 		t.Fatalf("Unexpected error message %s", diff.PrintWantGot(d))
 	}
+}
+
+func TestEmitCloudEvents(t *testing.T) {
+
+	object := &v1alpha1.Run{
+		ObjectMeta: metav1.ObjectMeta{
+			SelfLink: "/run/test1",
+		},
+		Status: v1alpha1.RunStatus{},
+	}
+	testcases := []struct {
+		name            string
+		data            map[string]string
+		wantEvents      []string
+		wantCloudEvents []string
+	}{{
+		name:            "without sink",
+		data:            map[string]string{},
+		wantEvents:      []string{},
+		wantCloudEvents: []string{},
+	}, {
+		name:            "with empty string sink",
+		data:            map[string]string{"default-cloud-events-sink": ""},
+		wantEvents:      []string{},
+		wantCloudEvents: []string{},
+	}, {
+		name:            "with sink",
+		data:            map[string]string{"default-cloud-events-sink": "http://mysink"},
+		wantEvents:      []string{},
+		wantCloudEvents: []string{`(?s)dev.tekton.event.run.started.v1.*test1`},
+	}}
+
+	for _, tc := range testcases {
+		// Setup the context and seed test data
+		ctx, _ := rtesting.SetupFakeContext(t)
+		ctx = WithClient(ctx, &FakeClientBehaviour{SendSuccessfully: true}, len(tc.wantCloudEvents))
+		fakeClient := Get(ctx).(FakeClient)
+
+		// Setup the config and add it to the context
+		defaults, _ := config.NewDefaultsFromMap(tc.data)
+		featureFlags, _ := config.NewFeatureFlagsFromMap(map[string]string{})
+		cfg := &config.Config{
+			Defaults:     defaults,
+			FeatureFlags: featureFlags,
+		}
+		ctx = config.ToContext(ctx, cfg)
+
+		recorder := controller.GetEventRecorder(ctx).(*record.FakeRecorder)
+		EmitCloudEvents(ctx, object)
+		if err := k8sevent.CheckEventsOrdered(t, recorder.Events, tc.name, tc.wantEvents); err != nil {
+			t.Fatalf(err.Error())
+		}
+		fakeClient.CheckCloudEventsUnordered(t, tc.name, tc.wantCloudEvents)
+	}
+}
+
+func TestEmitCloudEventsWhenConditionChange(t *testing.T) {
+	objectStatus := duckv1.Status{
+		Conditions: []apis.Condition{{
+			Type:   apis.ConditionSucceeded,
+			Status: corev1.ConditionUnknown,
+			Reason: v1beta1.PipelineRunReasonStarted.String(),
+		}},
+	}
+	object := &v1beta1.PipelineRun{
+		ObjectMeta: metav1.ObjectMeta{
+			SelfLink: "/pipelineruns/test1",
+		},
+		Status: v1beta1.PipelineRunStatus{Status: objectStatus},
+	}
+	after := &apis.Condition{
+		Type:    apis.ConditionSucceeded,
+		Status:  corev1.ConditionUnknown,
+		Message: "just starting",
+	}
+
+	data := map[string]string{"default-cloud-events-sink": "http://mysink"}
+	wantCloudEvents := []string{`(?s)dev.tekton.event.pipelinerun.started.v1.*test1`}
+
+	// Setup the context and seed test data
+	ctx, _ := rtesting.SetupFakeContext(t)
+	ctx = WithClient(ctx, &FakeClientBehaviour{SendSuccessfully: true}, len(wantCloudEvents))
+	fakeClient := Get(ctx).(FakeClient)
+
+	// Setup the config and add it to the context
+	defaults, _ := config.NewDefaultsFromMap(data)
+	featureFlags, _ := config.NewFeatureFlagsFromMap(map[string]string{})
+	cfg := &config.Config{
+		Defaults:     defaults,
+		FeatureFlags: featureFlags,
+	}
+	ctx = config.ToContext(ctx, cfg)
+
+	EmitCloudEventsWhenConditionChange(ctx, nil, after, object)
+	fakeClient.CheckCloudEventsUnordered(t, "with sink", wantCloudEvents)
 }
 
 func setupFakeContext(t *testing.T, behaviour FakeClientBehaviour, withClient bool, expectedEventCount int) context.Context {
