+++ pkg/reconciler/events/event_test.go
@@ -17,15 +17,12 @@ limitations under the License.
 package events
 
 import (
-	"errors"
 	"testing"
-	"time"
 
 	"github.com/tektoncd/pipeline/pkg/apis/config"
-	"github.com/tektoncd/pipeline/pkg/apis/pipeline/v1alpha1"
 	"github.com/tektoncd/pipeline/pkg/apis/pipeline/v1beta1"
 	"github.com/tektoncd/pipeline/pkg/reconciler/events/cloudevent"
-	eventstest "github.com/tektoncd/pipeline/test/events"
+	"github.com/tektoncd/pipeline/pkg/reconciler/events/k8sevent"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/tools/record"
@@ -34,147 +31,6 @@ import (
 	"knative.dev/pkg/controller"
 	rtesting "knative.dev/pkg/reconciler/testing"
 )
-
-func TestSendKubernetesEvents(t *testing.T) {
-	testcases := []struct {
-		name       string
-		before     *apis.Condition
-		after      *apis.Condition
-		wantEvents []string
-	}{{
-		name: "unknown to true with message",
-		before: &apis.Condition{
-			Type:   apis.ConditionSucceeded,
-			Status: corev1.ConditionUnknown,
-		},
-		after: &apis.Condition{
-			Type:    apis.ConditionSucceeded,
-			Status:  corev1.ConditionTrue,
-			Message: "all done",
-		},
-		wantEvents: []string{"Normal Succeeded all done"},
-	}, {
-		name: "true to true",
-		before: &apis.Condition{
-			Type:               apis.ConditionSucceeded,
-			Status:             corev1.ConditionTrue,
-			LastTransitionTime: apis.VolatileTime{Inner: metav1.NewTime(time.Now())},
-		},
-		after: &apis.Condition{
-			Type:               apis.ConditionSucceeded,
-			Status:             corev1.ConditionTrue,
-			LastTransitionTime: apis.VolatileTime{Inner: metav1.NewTime(time.Now().Add(5 * time.Minute))},
-		},
-		wantEvents: []string{},
-	}, {
-		name: "false to false",
-		before: &apis.Condition{
-			Type:   apis.ConditionSucceeded,
-			Status: corev1.ConditionFalse,
-		},
-		after: &apis.Condition{
-			Type:   apis.ConditionSucceeded,
-			Status: corev1.ConditionFalse,
-		},
-		wantEvents: []string{},
-	}, {
-		name: "unknown to unknown",
-		before: &apis.Condition{
-			Type:    apis.ConditionSucceeded,
-			Status:  corev1.ConditionUnknown,
-			Reason:  "",
-			Message: "",
-		},
-		after: &apis.Condition{
-			Type:    apis.ConditionSucceeded,
-			Status:  corev1.ConditionUnknown,
-			Reason:  "foo",
-			Message: "bar",
-		},
-		wantEvents: []string{"Normal foo bar"},
-	}, {
-		name:  "true to nil",
-		after: nil,
-		before: &apis.Condition{
-			Type:   apis.ConditionSucceeded,
-			Status: corev1.ConditionTrue,
-		},
-		wantEvents: []string{},
-	}, {
-		name:   "nil to true",
-		before: nil,
-		after: &apis.Condition{
-			Type:   apis.ConditionSucceeded,
-			Status: corev1.ConditionTrue,
-		},
-		wantEvents: []string{"Normal Succeeded "},
-	}, {
-		name:   "nil to unknown with message",
-		before: nil,
-		after: &apis.Condition{
-			Type:    apis.ConditionSucceeded,
-			Status:  corev1.ConditionUnknown,
-			Message: "just starting",
-		},
-		wantEvents: []string{"Normal Started "},
-	}, {
-		name: "unknown to false with message",
-		before: &apis.Condition{
-			Type:   apis.ConditionSucceeded,
-			Status: corev1.ConditionUnknown,
-		},
-		after: &apis.Condition{
-			Type:    apis.ConditionSucceeded,
-			Status:  corev1.ConditionFalse,
-			Message: "really bad",
-		},
-		wantEvents: []string{"Warning Failed really bad"},
-	}, {
-		name:   "nil to false",
-		before: nil,
-		after: &apis.Condition{
-			Type:   apis.ConditionSucceeded,
-			Status: corev1.ConditionFalse,
-		},
-		wantEvents: []string{"Warning Failed "},
-	}}
-
-	for _, ts := range testcases {
-		fr := record.NewFakeRecorder(1)
-		tr := &corev1.Pod{}
-		sendKubernetesEvents(fr, ts.before, ts.after, tr)
-		err := eventstest.CheckEventsOrdered(t, fr.Events, ts.name, ts.wantEvents)
-		if err != nil {
-			t.Errorf(err.Error())
-		}
-	}
-}
-
-func TestEmitError(t *testing.T) {
-	testcases := []struct {
-		name       string
-		err        error
-		wantEvents []string
-	}{{
-		name:       "with error",
-		err:        errors.New("something went wrong"),
-		wantEvents: []string{"Warning Error something went wrong"},
-	}, {
-		name:       "without error",
-		err:        nil,
-		wantEvents: []string{},
-	}}
-
-	for _, ts := range testcases {
-		fr := record.NewFakeRecorder(1)
-		tr := &corev1.Pod{}
-		EmitError(fr, ts.err, tr)
-		err := eventstest.CheckEventsOrdered(t, fr.Events, ts.name, ts.wantEvents)
-		if err != nil {
-			t.Errorf(err.Error())
-		}
-	}
-}
 
 func TestEmit(t *testing.T) {
 	objectStatus := duckv1.Status{
@@ -234,61 +90,7 @@ func TestEmit(t *testing.T) {
 
 		recorder := controller.GetEventRecorder(ctx).(*record.FakeRecorder)
 		Emit(ctx, nil, after, object)
-		if err := eventstest.CheckEventsOrdered(t, recorder.Events, tc.name, tc.wantEvents); err != nil {
-			t.Fatalf(err.Error())
-		}
-		fakeClient.CheckCloudEventsUnordered(t, tc.name, tc.wantCloudEvents)
-	}
-}
-
-func TestEmitCloudEvents(t *testing.T) {
-
-	object := &v1alpha1.Run{
-		ObjectMeta: metav1.ObjectMeta{
-			SelfLink: "/run/test1",
-		},
-		Status: v1alpha1.RunStatus{},
-	}
-	testcases := []struct {
-		name            string
-		data            map[string]string
-		wantEvents      []string
-		wantCloudEvents []string
-	}{{
-		name:            "without sink",
-		data:            map[string]string{},
-		wantEvents:      []string{},
-		wantCloudEvents: []string{},
-	}, {
-		name:            "with empty string sink",
-		data:            map[string]string{"default-cloud-events-sink": ""},
-		wantEvents:      []string{},
-		wantCloudEvents: []string{},
-	}, {
-		name:            "with sink",
-		data:            map[string]string{"default-cloud-events-sink": "http://mysink"},
-		wantEvents:      []string{},
-		wantCloudEvents: []string{`(?s)dev.tekton.event.run.started.v1.*test1`},
-	}}
-
-	for _, tc := range testcases {
-		// Setup the context and seed test data
-		ctx, _ := rtesting.SetupFakeContext(t)
-		ctx = cloudevent.WithClient(ctx, &cloudevent.FakeClientBehaviour{SendSuccessfully: true}, len(tc.wantCloudEvents))
-		fakeClient := cloudevent.Get(ctx).(cloudevent.FakeClient)
-
-		// Setup the config and add it to the context
-		defaults, _ := config.NewDefaultsFromMap(tc.data)
-		featureFlags, _ := config.NewFeatureFlagsFromMap(map[string]string{})
-		cfg := &config.Config{
-			Defaults:     defaults,
-			FeatureFlags: featureFlags,
-		}
-		ctx = config.ToContext(ctx, cfg)
-
-		recorder := controller.GetEventRecorder(ctx).(*record.FakeRecorder)
-		EmitCloudEvents(ctx, object)
-		if err := eventstest.CheckEventsOrdered(t, recorder.Events, tc.name, tc.wantEvents); err != nil {
+		if err := k8sevent.CheckEventsOrdered(t, recorder.Events, tc.name, tc.wantEvents); err != nil {
 			t.Fatalf(err.Error())
 		}
 		fakeClient.CheckCloudEventsUnordered(t, tc.name, tc.wantCloudEvents)
